FastAPI Backend/requirements.txt
fastapi
uvicorn
Pillow
requests
transformers
torch